requirements.txt
discord.py
python-dotenv
aiomysql
aiohttp